Added photo tab
Improved the previous UI. Added the photo tab. All UI in the photo tab works. It is possible to send an email of the photo. Script that takes a photo was not implemented. Thus what needs to be done: When "Take Photo" button is clicked it sends a request to the server. Then the server should run the python script to take the photo and save the photo in jpg format. The location of the taken photo relative to DexterApp folder should be returned to client as string. Then UI displays the photo and rest of functionality should work fine.